Remove irrelevant check from import
This change makes the code easier to read while removes a check
that would never be triggered.
After testing I verified that this error is only thrown when
a GPIO is actually used (for example by calling GPIO.setup(...)) not
when the module is imported.
(Note to self, do not trust the internet blindly...)